+++ BookList/Presentation/Presentation-Domain/Screen/Chat/ChatRoom/ChatRoomViewController.swift
@@ -5,6 +5,7 @@ final class ChatRoomViewController: UIViewController {
     @IBOutlet weak var tableView: UITableView!
 
     private var viewModel: ChatRoomViewModel!
+    private var dataSource: ChatRoomDataSource!
 
     private lazy var keyboardAccessoryView: KeyboardAccessoryView = {
         let view = KeyboardAccessoryView()
@@ -33,42 +34,11 @@ final class ChatRoomViewController: UIViewController {
     }
 
     private func setupTableView() {
+        dataSource = ChatRoomDataSource(viewModel: viewModel)
         tableView.register(MyMessageTableViewCell.xib(), forCellReuseIdentifier: MyMessageTableViewCell.resourceName)
-        tableView.delegate = self
-        tableView.dataSource = self
+        tableView.dataSource = dataSource
         tableView.estimatedRowHeight = 80
         tableView.rowHeight = UITableView.automaticDimension
-    }
-}
-
-extension ChatRoomViewController: UITableViewDelegate {
-
-}
-
-extension ChatRoomViewController: UITableViewDataSource {
-
-    func tableView(
-        _ tableView: UITableView,
-        numberOfRowsInSection section: Int
-    ) -> Int {
-        viewModel.messages.count
-    }
-
-    func tableView(
-        _ tableView: UITableView,
-        cellForRowAt indexPath: IndexPath
-    ) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(
-            withIdentifier: MyMessageTableViewCell.resourceName,
-            for: indexPath
-        )
-
-        if let chatRoomCell = cell as? MyMessageTableViewCell {
-            chatRoomCell.backgroundColor = .clear
-            chatRoomCell.userMessageTextView.text = viewModel.messages[indexPath.row]
-        }
-
-        return cell
     }
 }
